+++ catlearn/fingerprint/convoluted.py
@@ -1,10 +1,10 @@
 """Slab adsorbate convoluted fingerprint functions for machine learning."""
 import numpy as np
-from ase.atoms import string2symbols
 from ase.data import ground_state_magnetic_moments as gs_magmom
-from ase.data import atomic_numbers
-from .periodic_table_data import list_mendeleev_params, default_params
-from .base import BaseGenerator
+
+from catlearn.featurize.periodic_table_data import (list_mendeleev_params,
+                                                    default_params)
+from catlearn.featurize.base import BaseGenerator
 
 default_convoluted_fingerprinters = ['conv_bulk',
                                      'conv_term']