+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # dev
 
+# Version 0.5.0 (August 2018)
+
+-   Restructure of fingerprint module
+-   Pandas DataFrame getter in FeatureGenerator
 -   CatMAP API using ASE database.
--   New surrogate model module.
+-   New active learning module.
 -   Small fixes in adsorbate fingerprinter.
 
 # Version 0.4.4 (August 2018)